requirements.txt
streamlit
openai-whisper
torch
spacy
yt-dlp
transformers